wacom-raw: Fix an impossible to hit overflow (#3082)
Spotted by Coverity.

+++ plugins/wacom-raw/fu-wacom-aes-device.c
@@ -175,10 +175,10 @@ fu_wacom_aes_device_write_block (FuWacomAesDevice *self,
 				 guint32 idx,
 				 guint32 address,
 				 const guint8 *data,
-				 guint16 datasz,
+				 gsize datasz,
 				 GError **error)
 {
-	guint blocksz = fu_wacom_device_get_block_sz (FU_WACOM_DEVICE (self));
+	gsize blocksz = fu_wacom_device_get_block_sz (FU_WACOM_DEVICE (self));
 	FuWacomRawRequest req = {
 		.cmd = FU_WACOM_RAW_BL_CMD_WRITE_FLASH,
 		.echo = (guint8) idx + 1,
@@ -194,7 +194,7 @@ fu_wacom_aes_device_write_block (FuWacomAesDevice *self,
 			     G_IO_ERROR,
 			     G_IO_ERROR_FAILED,
 			     "block size 0x%x != 0x%x untested",
-			     datasz, (guint) blocksz);
+			     (guint) datasz, (guint) blocksz);
 		return FALSE;
 	}
 	memcpy (&req.data, data, datasz);
